Admin::ApiController: copy this file into parent/root application
This file is essentially the configuration that runs the admin/documents#index view. If you want to add or remove facet fields in the admin search view, you need to make changes to this file.

Moving the file into the parent/root application allows local modification of the facets.

Addresses #32